requirements.txt
python-telegram-bot==20.7
requests==2.31.0 
ruff==0.7.0
pytz
dateutil